requirements.txt
requests~=2.26.0
snowflake-connector-python==2.4.0